Mods to vcf classes for the VCFComparator.  Set stringency of Sam2USeq SAMReaders to silent.

+++ trunk/Source/edu/utah/seq/useq/data/RegionScoreText.java
@@ -77,8 +77,8 @@ public class RegionScoreText extends RegionScore{
 		this.text = text;
 	}
 	/**Assumes interbase coordinates.*/
-	public static int countBases(HashMap<String,RegionScoreText[]> regions){
-		int total = 0;
+	public static long countBases(HashMap<String,RegionScoreText[]> regions){
+		long total = 0;
 		for (RegionScoreText[] r : regions.values()){
 			for (int i=0; i< r.length; i++){
 				total += r[i].getLength();
